+++ pons/_provider.py
@@ -1,15 +1,13 @@
 from abc import ABC, abstractmethod
 from contextlib import asynccontextmanager
 from http import HTTPStatus
-from typing import Any, AsyncIterator, Dict, Optional, Union, cast, Iterable, Mapping
+from typing import Any, AsyncIterator, Dict, Optional, Union, cast, Iterable, Mapping, Tuple
 
 import httpx
 
 
-# TODO: currently mypy does not support recursive type aliases,
-# make it recursive when it's possible.
-# See https://github.com/python/mypy/issues/731
-JSON = Union[bool, int, float, str, None, Iterable[Any], Mapping[str, Any]]
+# TODO: the doc entry had to be written manually for this type because of Sphinx limitations.
+JSON = Union[bool, int, float, str, None, Iterable["JSON"], Mapping[str, "JSON"]]
 
 
 class Provider(ABC):
@@ -67,13 +65,33 @@ class ProviderSession(ABC):
     """
 
     @abstractmethod
-    async def rpc(self, method: str, *args: JSON) -> Any:
+    async def rpc(self, method: str, *args: JSON) -> JSON:
         """
         Calls the given RPC method with the already json-ified arguments.
         """
         ...
 
+    async def rpc_and_pin(self, method: str, *args: JSON) -> Tuple[JSON, Tuple[int, ...]]:
+        """
+        Calls the given RPC method and returns the path to the provider it succeded on.
+        This method will be typically overriden by multi-provider implementations.
+        """
+        return await self.rpc(method, *args), ()
+
+    async def rpc_at_pin(self, path: Tuple[int, ...], method: str, *args: JSON) -> JSON:
+        """
+        Calls the given RPC method at the provider by the given path
+        (obtained previously from ``rpc_and_pin()``).
+        This method will be typically overriden by multi-provider implementations.
+        """
+        if path != ():
+            raise ValueError(f"Unexpected provider path: {path}")
+        return await self.rpc(method, *args)
+
     async def rpc_dict(self, method: str, *args: JSON) -> Optional[ResponseDict]:
+        """
+        Calls the given RPC method expecting to get a dictionary (or ``null``) in response.
+        """
         result = await self.rpc(method, *args)
         if result is None:
             return None
@@ -101,7 +119,7 @@ class RPCError(Exception):
     """
 
     @classmethod
-    def from_json(cls, response: Dict[str, JSON]) -> "RPCError":
+    def from_json(cls, response: JSON) -> "RPCError":
         error = ResponseDict(response)
         if "data" in error:
             data = error["data"]
@@ -158,12 +176,13 @@ class HTTPSession(ProviderSession):
         if response.status_code != HTTPStatus.OK:
             raise RPCError(response.status_code, response.content.decode())
 
-        response_json = cast(JSON, response.json())
-        if not isinstance(response_json, dict):
+        response_json = response.json()
+        if not isinstance(response_json, Mapping):
             raise UnexpectedResponse(f"RPC response must be a dictionary, got: {response_json}")
+        # Assuming that the HTTP client knows what it's doing, and gives us a valid JSON dict
+        response_json = cast(Mapping[str, JSON], response_json)
         if "error" in response_json:
             raise RPCError.from_json(response_json["error"])
         if "result" not in response_json:
             raise UnexpectedResponse(f"`result` is not present in the response: {response_json}")
-        # TODO: see the TODO above; when JSON is recursive, this cast won't be necessary.
-        return cast(JSON, response_json["result"])
+        return response_json["result"]